+++ src/components/myPage/NotiSetting.tsx
@@ -3,70 +3,56 @@
 import ToggleButton from '../common/ToggleButton';
 import Button from '../common/Button';
 import { useState } from 'react';
-import Toast from '../common/Toast';
 import MoreRightIcon from '../common/icons/MoreRight';
-
+import { useToastStore } from '@/stores/toast';
+import { getToday } from '@/utils/date';
 
 export default function NotiSetting() {
-  const [serviceToast, setServiceToast] = useState(true);
-  const [showServiceToast, setShowServiceToast] = useState(false);
-  const [toastMessage, setToastMessage] = useState('');
-  const [eventToast, setEventToast] = useState(false);
-  const [showEventToast, setShowEventToast] = useState(false);
+  const setToast = useToastStore.use.setToast();
+  const [serviceOn, setServiceOn] = useState(true);
+  const [eventOn, setEventOn] = useState(false);
 
   // 주요 서비스 알림
   const handleToggleService = () => {
-    if (serviceToast) {
-      // 서비스 알림을 끌 때
-      setToastMessage(
-        '주요 서비스 알림을 해제하여\n가격 할인 알람을 받을 수 없어요!',
-      );
-    } else {
-      // 서비스 알림을 켤 때
-      setToastMessage(
-        '주요 서비스 알림을 켜서\n가격 할인 알림을 받을 수 있어요!',
-      );
-    }
-    setShowServiceToast(true);
-    setTimeout(() => setShowServiceToast(false), 3000);
-    setServiceToast(!serviceToast);
+    setToast(
+      serviceOn
+        ? '주요 서비스 알림을 해제하여\n가격 할인 알람을 받을 수 없어요!'
+        : '주요 서비스 알림을 켜서\n가격 할인 알림을 받을 수 있어요!',
+      3000,
+    );
+    setServiceOn((prev) => !prev);
   };
 
   // 마켓팅 정보 수신 동의
   const handleToggleEvent = () => {
-    if (eventToast) {
-      // 서비스 알림을 끌 때
-      setToastMessage('2024년 11월 5일, 마케팅 정보 수신 동의를 철회했어요.');
-    } else {
-      // 서비스 알림을 켤 때
-      setToastMessage('2024년 11월 5일, 마케팅 정보 수신 동의를 유지해요.');
-    }
-    setShowEventToast(true);
-    setTimeout(() => setShowEventToast(false), 3000);
-    setEventToast(!eventToast);
+    const today = getToday();
+    setToast(
+      eventOn
+        ? `${today}, 마케팅 정보 수신 동의를 철회했어요.`
+        : `${today}, 마케팅 정보 수신 동의를 유지해요.`,
+      3000,
+    );
+    setEventOn((prev) => !prev);
   };
 
   return (
-    <div className="border-b border-border relative">
-      <div className="px-4 py-3.5 border-b bg-bar border-border">
+    <div className="relative border-b border-border">
+      <div className="border-b border-border bg-bar px-4 py-3.5">
         <span className="text-sm text-gray">알림 설정</span>
       </div>
       <div className="pb-4">
         <div className="flex items-center justify-between px-4 py-2">
           <h2 className="text-base text-dark-gray">주요 서비스 알림</h2>
-          <ToggleButton
-            isOn={serviceToast}
-            toggleHandler={handleToggleService}
-          />
+          <ToggleButton isOn={serviceOn} toggleHandler={handleToggleService} />
         </div>
-        <p className="text-sm text-gray-text px-4 w-96">
+        <p className="w-96 px-4 text-sm text-gray-text">
           역대 최저가, 평균 대비 하락세, 설정된 n% 도달 시 알림을 받을 수
           있습니다.
         </p>
       </div>
       <div className="flex items-center justify-between px-4">
         <h3 className="text-base text-dark-gray">광고성 알림</h3>
-        <ToggleButton isOn={eventToast} toggleHandler={handleToggleEvent} />
+        <ToggleButton isOn={eventOn} toggleHandler={handleToggleEvent} />
       </div>
       <div className="flex items-center justify-between px-4 pb-5 pt-2">
         <p className="pb-1 text-sm text-gray-text">
@@ -76,16 +62,6 @@ export default function NotiSetting() {
           <MoreRightIcon />
         </Button>
       </div>
-      {showServiceToast && (
-        <div className="fixed bottom-5 left-1/2 transform -translate-x-1/2 text-sm">
-          <Toast text={toastMessage} />
-        </div>
-      )}
-      {showEventToast && (
-        <div className="fixed bottom-5 left-1/2 transform -translate-x-1/2 text-sm">
-          <Toast text={toastMessage} position="flex" />
-        </div>
-      )}
     </div>
   );
 }
